Refactor invetory
