+++ finish/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -13,9 +13,14 @@
 // tag::InventoryManager[]
 package io.openliberty.guides.inventory;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Properties;
 
 import io.openliberty.guides.inventory.model.InventoryList;
+import io.openliberty.guides.inventory.model.SystemData;
+
 import javax.enterprise.context.ApplicationScoped;
 import org.eclipse.microprofile.metrics.annotation.Counted;
 import org.eclipse.microprofile.metrics.annotation.Timed;
@@ -27,32 +32,36 @@ import org.eclipse.microprofile.metrics.MetricUnits;
 // end::ApplicationScoped[]
 public class InventoryManager {
 
-  private InventoryList invList = new InventoryList();
+  private List<SystemData> systems = Collections.synchronizedList(new ArrayList<>());
   private InventoryUtils invUtils = new InventoryUtils();
 
   @Timed(name = "inventoryPropertiesRequestTime", absolute = true,
     description = "Time needed to get the properties of a system from the given hostname")
   public Properties get(String hostname) {
-    Properties properties = invUtils.getProperties(hostname);
-
-    if (properties != null) {
-      invList.addToInventoryList(hostname, properties);
-    }
-
-    return properties;
-
+    return invUtils.getProperties(hostname);
   }
+
+  public void add(String hostname, Properties systemProps) {
+    Properties props = new Properties();
+    props.setProperty("os.name", systemProps.getProperty("os.name"));
+    props.setProperty("user.name", systemProps.getProperty("user.name"));
+
+    SystemData host = new SystemData(hostname, props);
+    if (!systems.contains(host))
+      systems.add(host);
+  }
+
 
   @Counted(name = "inventoryAccessCount", absolute = true, monotonic = true,
     description = "Number of times the list of systems method is requested")
   public InventoryList list() {
-    return invList;
+    return new InventoryList(systems);
   }
 
   @Gauge(unit = MetricUnits.NONE, name = "inventorySizeGuage", absolute = true,
     description = "Number of systems in the inventory")
   public int getTotal() {
-    return invList.getSystems().size();
+    return systems.size();
   }
 
 }
